// ui-source/src/lib/utils/types.ts
export interface PlayersData {
	duty: boolean;
	radio: number;
	callsign: string;
	firstname: string;
	lastname: string;
	citizenid: string;
	vehicle: object;
	assignment: object;
}

export interface MyData {
	duty: boolean;
	firstname: string;
	lastname: string;
	citizenid: string;
	callsign:string
}